+++ parcels/interaction/base_neighbor.py
@@ -7,13 +7,16 @@ from timeit import timeit
 class BaseNeighborSearch(ABC):
     name = "unknown"
 
-    def __init__(self, interaction_distance, interaction_depth, values=None,
-                 max_depth=100000):
-        self._values = values
+    def __init__(self, interaction_distance, interaction_depth, max_depth=100000):
+#         self._values = values
         self.interaction_depth = interaction_depth
         self.interaction_distance = interaction_distance
-        self.inter_dist = np.array([self.interaction_distance, self.interaction_distance, self.interaction_depth]).reshape(3, 1)
+        self.inter_dist = np.array(
+            [self.interaction_distance, self.interaction_distance,
+             self.interaction_depth]).reshape(3, 1)
         self.max_depth = max_depth
+        self._values = None
+        self._active_mask = None
 
     def find_neighbors_by_idx(self, particle_idx):
         '''Find neighbors with particle_id.'''
@@ -24,11 +27,25 @@ class BaseNeighborSearch(ABC):
     def find_neighbors_by_coor(self, coor):
         raise NotImplementedError
 
-    def update_values(self, new_values):
+    def update_values(self, new_values, new_active_mask=None):
         self._values = new_values
+        self._active_mask = new_active_mask
 
-    def rebuild(self, values):
-        self._values = values
+    def rebuild(self, values, active_mask=-1):
+        if values is not None:
+            self._values = values
+        if active_mask is None:
+            self._active_mask = np.arange(self._values.shape[1])
+        if active_mask != -1:
+            self._active_mask = active_mask
+#         if self._active_mask is None:
+#             self._active_idx = None
+
+    @property
+    def active_idx(self):
+        if self._active_mask is None:
+            return None
+        return np.where(self._active_maske)[0]
 
     @classmethod
     def benchmark(cls, max_n_particles=1000, density=1):
